local-starter/pyproject.toml
[project]
name = "lfi-workflow-template"
version = "0.1.0"
description = "Two step local inocation workflow template"
authors = [
    { name = "Resonate HQ", email = "contact@resonatehq.io" }
]
dependencies = [
    "resonate-sdk>=0.5.6",
]
readme = "README.md"
requires-python = "== 3.12.*"

[project.scripts]
    "app" = "src.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
